requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.114.0
frozenlist==1.4.1
h11==0.14.0
idna==3.8
multidict==6.1.0
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
yarl==1.11.1
